requirements.txt
fastapi
uvicorn
openai
python-dotenv
httpx
gunicorn